move model specific files to `models`
