+++ src/ff16_strategy.cpp
@@ -1,9 +1,7 @@
-#include <plant/ff16_strategy.h>
 #include <plant/uniroot.h>
 #include <plant/qag.h>
-#include <plant/environment.h>
 #include <plant/assimilation.h>
-#include <plant/strategy.h>
+#include <plant/models/ff16_strategy.h>
 #include <RcppCommon.h> // NA_REAL
 
 namespace plant {
